+++ webapp-frontend/src/app/services/caas-ingress/caas-ingress.service.spec.ts
@@ -7,9 +7,9 @@
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ * 
  *      http://www.apache.org/licenses/LICENSE-2.0
- *
+ * 
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -17,18 +17,15 @@
  * limitations under the License.
  * ========================LICENSE_END===================================
  */
-import { Component, OnInit } from '@angular/core';
+import { TestBed } from '@angular/core/testing';
 
-@Component({
-  selector: 'rd-control',
-  templateUrl: './control.component.html',
-  styleUrls: ['./control.component.scss']
-})
-export class ControlComponent implements OnInit {
+import { CaasIngressService } from './caas-ingress.service';
 
-  constructor() { }
+describe('CaasIngressService', () => {
+  beforeEach(() => TestBed.configureTestingModule({}));
 
-  ngOnInit() {
-  }
-
-}
+  it('should be created', () => {
+    const service: CaasIngressService = TestBed.get(CaasIngressService);
+    expect(service).toBeTruthy();
+  });
+});